* Create new csv2ds program that converts csv files to dataseries.
check-data/csv2ds-1.*: testing data for regression test

csv2ds.cpp: New program

run-check-csv2ds.sh: test program

ExtentType.hpp: Add example of an ExtentType to the documentation

GeneralField.hpp: convert BUG/FEATURE to TODO, add a set(string) function to
  GeneralValues for use by csv2ds

GeneralField.cpp: Move all the interpretations of "true" and "false" for 
  boolean strings to the top.  Implement conversion from variable32 to boolean.
  Change all the set(GeneralValue) functions to use val* rather than requiring
  the value and field type to be identical.

process/CMakeLists.txt: add new program

wcweb2ds.cpp: fix up coding bits

tests/CMakeLists.txt: add new script test.

+++ src/process/wcweb2ds.cpp
@@ -91,11 +91,9 @@ struct record {
     uint8_t method, status, type, server;
 };
 
-unsigned nrecords;
+uint64_t nrecords;
 
-int
-main(int argc,char *argv[])
-{
+int main(int argc,char *argv[]) {
     commonPackingArgs packing_args;
     getPackingArgs(&argc,argv,&packing_args);
